run: stop using unshare and chroot.
This introduces the b5-enter command, which does the exact logic that we
want to perform in order to enter in an isolated container. The command
acts like a mix of unshare, nsenter, chroot, mount... and does all of the
low-level initialisation with semantics that we define, so we don't end
up surprised by whatever coreutils implementation is available locally.

b5-enter can be used standalone, but is otherwise the main driver behind
`b5 run` to enter image snapshots on their own terms.

Change-Id: I759947ba71f451012076b729c4f5dae84fbb772f