b5-enter needs to ensure target directories exist before mounting
During image building, we will often mount to places like
/src/github.com/foo/bar inside the image, which almost never actually
exist in the image. With this patch, b5-enter creates these directories
so that it can mount to them.

This is probably not a good idea, because it actually creates
directories in the floor - it would be better to first mount /src
to a tmpfs and then create directories in the tmpfs that can
hold the source mount. I will come back to this later, but I'm
punting it for now in order to get unblocked on image building.

Change-Id: I01243a0d290fca45cc4751f336486d85a1729dd5