b5-enter: do not try to change [ug]ids when no flag is passed
This helps with testing, and is a step towards allowing
b5-enter / exec argv... to be run unprivileged.

setre[ug]id has the very nice property of not doing anything when their
parameters are -1. As for setgroups, calling it with a size of 0 does
*not* drop all additional groups (you always have your main gid as one),
so we reuse that value to signify no change.

Change-Id: Ieab7d6c89839a3aba87f8fd89e06fff5e26f152b