b5-enter: add more elaborate signal handling
This fixes an issue where we couldn't Ctrl-C b5-enter if the inner process
was stuck. That happened because we ignored the signals using SIG_IGN,
which gets inherited by children and crosses exec() boundaries.

This new code is much better because not only does it avoids these
issues, but is also actively ignores any user-sent signal while still
doing the right think w.r.t kernel-sent SIGSEGVs, SIGBUSes, and other
signals that are triggered by the execution of illegal opcodes or
syscalls.

Change-Id: I6fb8c923921bdfcb094caa0101a7de324ac68f3c