b5-enter: arrange for recursive use of user namespaces
man 2 unshare documents that unshare may fail with EPERM if done after a chroot:

    EPERM (since Linux 3.9)
        CLONE_NEWUSER was specified in flags and the caller is in a chroot
        environment  (i.e., the  caller's  root directory does not match the
        root directory of the mount namespace in which it resides).

This is unfortunately always true when employing chroot. The proper way
to handle this is to use pivot_root, which changes the view world for every
process in the mount namespace. This allows us to nest user namespaces, but
we introduce one limitation.

Because pivot_root is really dangerous, we insist that it must be done
in a mount namespace. This means that when running b5-enter -U without
-m, you won't be able to nest more user namespaces.

Change-Id: Ifd743c9cad515b43f0665d75825a53b542a1467a