chroot: resolve root and default to /
One of the immediate consequences of bad55ca is that commands like
`bst -r /` completely break. This is because we make some preparations
for pivot_root that only make sense for any directory other than /, and
pivot_root itself insists that the new root must not be "/".

The way to fix this for / and any path that end up being / like //, /./, /../,
... is to simply call realpath on the root and use the result as a basis
for our comparison.

Change-Id: Ia13778b5a9ed9fe93f6acdac70e1f0fe33b5b30c