mount: support read-only bind mounts
It turns out that you cannot make a read-only bind-mount in one pass: it
is necessary to remount the existing entry read-only.

There are caveats in doing that in a user namespace. Namely, if the
source is in a tmpfs that isn't owned by you, and you haven't specified
neither `nosuid` nor `nodev` while mounting the read-only bind-mount,
then it will fail with EPERM. This, however, is not a bst problem: if
a user really, really wants to mount read-only a host tmpfs in a
spacetime, they'll have to call bst with `sudo unshare -m` and do the
privileged mount. It is not bst's job to circumvent the kernel's security
policies.

Change-Id: I1e0a9009a89c801e3e3a360acde5d64f2236ff55