time,ns: reset all supported clocks to 0
This change does not make any assumption about the kind of clocks
supported in time namespaces -- it tries to reset all of them, and
ignores clocks it can't offset.

Disappointingly, at the time of writing, CLOCK_REALTIME is not supported
in time namespaces, but its support is not something that can be ruled
out in the future. In respect to this, this change is
forward-compatible.