mount,devtmpfs: implement fake devtmpfs mounts
This commit introduces a bst_devtmpfs mount fstype, which causes a tmpfs
to be mounted on the destination, then filled with a minimal set of
non-problematic vital devices (e.g. /dev/null).

The implementation first tries to mknod the relevant devices. If this is
not possible (for instance, EPERM in user namespaces), it falls back to
mounting individual devices from the host /dev.

If a user asks bst to mount a devtmpfs with --mount, the fstype is
automatically translated to bst_devtmpfs. Because sometimes it's not desirable
to have this fake devtmpfs, a --no-fake-devtmpfs flag is introduced to
disable the behaviour.