Revamp --share options, add --share-<ns>=<file> to invoke setns()
This commit refactors the way --share options are processed.  There is
now a "--share-<ns>" option for each of the eight namespace types (now
uniformly named "cgroup", "ipc", "mnt", "net", "pid", "time", "user",
and "uts").  If not specified, the namespace is unshared.  If
specified with no value, the namespace is shared with parent.  If
specified with a value, the value is a file that we open, and then
share the namespace in the file.

This commit also provides a "--share-all" option, which causes all
supported namespaces to be shared with parent.  If an optional value
is provided, then we share the namespace file in the directory with
a matching name (/the/directory/cgroup, /the/directory/ipc, etc).
So, you can `bst --share-all=/proc/[pid]/ns` to join all namespaces
of the given process.  So elegant!

In this commit, we continue to support --share=ns[,ns...] (deprecated)
for backward compatibility with barney.  We will remove this support
once barney is upgraded.