Fix bug where inner could still run after failure in outer_helper
Before this change, when outer dies for some reason, it was possible
for inner to run for a short time before the parent died (due to outer
dying) which would kill inner (due to PR_SET_PDEATHSIG):

   root@localhost:~/src/github.com/aristanetworks/bst# bst id
   bst: burn(uid_map): write: Operation not permitted
   bst: outer_helper_sendpid_and_wait: helper exit status 1
   root@localhost:~/src/github.com/aristanetworks/bst# uid=65534(nobody) gid=65534(nogroup) groups=65534(nogroup)

Note the output of "id" showing up there.  Bad!

This fixes the bug by making sure that outer actually sent us the
"okay" message on our pipe, rather than closing the pipe for some
other reason (like, because of dying).