init: make builtin init dumpable
This lets us use tools like nsenter (or ourselves) to join namespaces
defined in /proc/pid/ns, where pid is the pid of the builtin init.
Previously, that directory would be owned by root, and we wouldn't be
able to open anything in there unless privileged. Setting the dumpable
flag grants us back ownership of the proc directory.