caps: use finer-grained on-demand capabilities
This is the correct thing to do w.r.t. capabilities -- the whole bst
codebase doesn't need to run with cap_sys_admin or cap_sys_ptrace. This
reduced the attack surface and already does the right thing for some of
the cli usage: this made apparent the fact that using --uid without
changing --worktree to somewhere where that uid has execute permission
down the chain is just wrong.