creds: switch credentials after setting up mounts
This simplifies the capability management of post-unshare code, and
exposes more clearly the boundary between userns-privileged operations
and post-setuid operations.

For instance, we previously executed the builtin bst init with uid and
gid set to whatever was passed in --uid and --gid. Now, init is always
run as the netns root user.