init: allow --init when not unsharing pidns
Specifying an init process when not unsharing a new pid namespace is
problematic, because we can't magically make the new process PID 1.

We have the choice of making --init invalid in such cases, but perhaps a more
sensible thing to do is to make that init process a subreaper. This has
some nice but odd properties -- in particular, bst can be used in this
manner to "undaemonize" processes while staying in the original PID
namespace.