cli: add new "--limit-<XXX>" options
To give the user more control over the bst process, add three new options for
runing `setrlimit` before exec.

    --limit-core: RLIMIT_CORE
    --limit-nofile: RLIMIT_NOFILE
    --limit-nproc: RLIMIT_NPROC

Each option requires an argument which is parsed as a colon-separated tuple of
hard, soft resource limits.  Each of hard and soft limits are optional, and
when one is not specified, it will use the current value.  The only exception
is for the special value ":" which will use the current hard limit, and set
the soft limit to the hard limit value.

    --limit-nofile=100 => setrlimit(RLIMIT_NOFILE, (100, 100))
    --limit-nofile=100:50 => setrlimit(RLIMIT_NOFILE, (100, 50))
    --limit-nofile=100: => setrlimit(RLIMIT_NOFILE, (100, cur))
    --limit-nofile=:100 => setrlimit(RLIMIT_NOFILE, (cur, 100))
    --limit-nofile=: => setrlimit(RLIMIT_NOFILE, (cur, hard))

Added test cases to bst.t and also added a unit test for the parsing
code.

Usage strings:
    --limit-core (<value>)|(<hard>:<soft>)
                            Set RLIMIT_CORE to the provided value(s).
    --limit-nofile (<value>)|(<hard>:<soft>)
                            Set RLIMIT_NOFILE to the provided value(s).
    --limit-nproc (<value>)|(<hard>:<soft>)
                            Set RLIMIT_NPROC to the provided value(s).