misc: clean up baseline -Wconversion warnings
net.c was a particular pain because of the NLA_HDRLEN macro being
internally incorrect, causing the warnings to be thrown wherever it
was used.  For those instances, I've just disabled the warning in the
relevant code sections.

net_if_rename() is populating the `struct ifinfomsg` `ifi_index`
value, so it needs to be a signed int value; changed the function
signature to `int link`.