limits: define resource constants ourselves
We can't really rely on whatever definitions are present on the "host",
given that the host might be a container, and the running kernel might
be different. This, in particular, is visible in the clashing
definitions of rlimit constants between sys/resource.h and
asm/resource.h.

To avoid this, we instead define the constants we support ourselves.
These values are guaranteed to be stable between platforms anyway,
so we can at runtime determine which values are supported by checking
for EINVAL.