init: don't put main child into its own process group
We lose the ability to kill the main process subtree, but it's fine
considering that init terminates when it reaps the main child, which
kills the rest of the pid namespace.

This change was motivated by bst rendering processes unkillable if they
were put in the background. `fg` would end up putting bst back in the
background rather than the new process group.