mount: make bst_devtmpfs work on /dev
Previously, bst would give up trying to setup the fake devtmpfs mount if
the target was the host /dev. This was because setting up the fake
devtmpfs involved bind-mounting devices from the original /dev.

This commit makes this problem disappear by preparing said mount on /tmp
if we're trying to mount over the host /dev.