Revert "net: use deterministic MACs"
This reverts (part of) commit f89740058f50508282ba5a6b5d1fbc73807cf8c7.

Trying to provide a deterministic default for MACs ends up biting us,
since we end up setting the same value for concurrent spacetimes. In
theory, it'd be possible to move the MAC assignation in the inner netns
but at this point it'd be more work than necessary, and the kernel is
perfectly capable of providing reasonable defaults anyway.