pty: fix signal delivery
We were only getting the first signal since pselect unblocks all
signals after it is called instead of just restoring the mask.

Removed the use of pselect because it was redundant.  We were
already blocking the signals.

Also removed some of the hacks that were in place to work around
missing SIGCHLD.

Finally, added correct setting of the term window size on parent setup.