misc: fix formatting
This makes the code format a bit more consistent with the rest of the
codebase. It also fixes functions with unspecified argument lists, as
these do not mean the same thing as (void) -- the former can be called
with any number of argument, while the latter can be called with exactly
zero.