tty: ensure we send the right VEOF control character
I noticed that bst would hang when the inner shell changed its eof
control character. This is because we assumed that the inner tty would
simply keep accepting EOT as end of input.

This does not solve this class of problems, however; it's possible that
an inner process can still hang itself by racing a change of the VEOF
character with the outer end of input. This is very unlikely to happen
however, and only causes the inner process to stop functioning.

For the common, sane case, probing the termios before sending the right
VEOF is the correct thing to do.