main: check for empty or missing seconds value as parsed from the command line
While trying to use the seconds value as provided by the --time command line
option, we need to check if the user did indeed pass any value at all or if the
passed value is an empty string. This change adds those two validations. The
same validations are also already done for the nanoseconds value.

Signed-off-by: Ani Sinha <ani@anisinha.ca>