cleanup: use CLOCK_TAI + 1 directly as MAX_CLOCK
At present MAX_CLOCK is defined to be a numeric value equal to CLOCK_TAI.
It is much cleaner to directly use MAX_TAI + 1 as the MAX_CLOCK value.
This way, MAX_CLOCK truly represents the maximum value of the valid clock
sources one can have. This also makes array initializations and iterations
cleaner.
It is to be noted that MAX_CLOCKS is also defined in the linux kernel header
and its value is 14. However, it is not exposed to the userland header. If it
had been exposed, it might have been even cleaner to use MAX_CLOCKS directly.

Signed-off-by: Ani Sinha <ani@anisinha.ca>