outer_helper_spawn: fail pid burn silently when all limits are non-critical
When running in an environment with legacy cgroups, the cgroups-related
parts of outer_helper_spawn aren't applicable. When a user invokes bst
with --try-limit, the code needs to fail softly, as if no limit had been
specified at all. We handle other cases of this earlier on the code, but
we may (apparently) reach as far as burning the child pid into
cgroup.procs before legacy cgroups balks, so we add another soft-failure
for that case here.