cgroup: add systemd driver
This allow things to _just work_ on systems managed by systemd. It's
just a bad idea to not try to integrate with it on these systems, as
systemd will try to own the entire cgroup hierarchy. Not integrating
with it was one of the primary reasons why the previous cgroup
implementation had many pitfalls and was hard to use correctly.