env: fix environment variable parsing
Environment variables are fairly free-form; all characters are allowed
except for '\0', so FOO-BAR=baz should work.

This leaves callers in a bit of a conundrum if they don't know (and
don't particularly trust) program names in advance. To make things
safer, and more explicit, a second -- can now be added to the
environment variable list.