cgroup,setup: make CGROUP_PATH optional
Some users of bst are still stuck on cgroups v1, and 588d370
("cgroup,setup: set CGROUP_PATH in setup program with full path to
current cgroup") broke the basic, non-cgroup use of bst.

This commit makes the failure to determine the current cgroup path
non-fatal; the setup program has the responsibility to fail if it needs
it.