tty: handle SIGHUP when allocating a tty
When a session is hung up and a tty is allocated SIGHUP will be sent
to all processes in the session group.  Since we were ignoring SIGHUP
this led to bst processes hanging around.

Add signal handling for SIGHUP to close the tty master.  This will lead
to cleanup of the child session.