add an option to skip draining on tty close (#105)
* add an option to skip draining on tty close

When ssh connections to buses are terminated bst is blocking while
draining the pty.  The signal handler does to run in this case and
the process hangs around until the bus is rebooted.

Added an option to control whether or not the pty is drained when
stdin is closed.  The b5c bst invocation will pass --tty=-drain.