cgroup,systemd: fix race in message dequeue waiting for cgroup to exist
The old code would assume that the very next message in the dbus queue
was the JobRemoved signal, signalling that the cgroup was created by
systemd and ready to be used.

Unfortunately, in the real world, other kinds of messages may be
inserted before getting to JobRemoved, and the old code would almost always
move on to assuming the job completed even if it didn't actually receive
the signal. This caused issues where running bst would fail due to the
cgroup directory not yet being present in the filesystem.

This commit fixes this situation by properly dequeueing the message queue
until no messages remain, and stopping once JobRemoved is found.