pdeathsig: use pidfd for parent liveness checks
The older way of opening a pipe whose write end would remain in the
parent and get closed was awkward and left quite a few sharp edges
that caused bugs previously (62f5ae9, 33d3f27).

This new approach is more purposeful in actually checking the liveness
of the parent rather than relying on a side effect.