pdeathsig: stop assuming getting sigkilled kills us
In the second user of sig_setpdeathsig, the current process is pid 1 of
its own pid namespace. This has the property of setting the disposition
for every signal to SIG_IGN, even SIGKILL.

This is very unfortunate for sig_setpdeathsig that assumes
kill(getpid(), signo) will promptly self-terminate the process if signo
is terminal and uncatcheable.

This commit addresses this corner case by explicitly calling exit if
such a situation is detected.