pdeathsig: fall back to pipe trick instead of proc comparison
In some scenarios we can't get to /proc/<ppid> because the parent is
outside of the child's pid namespace, and therefore is ppid is 0, making
even the fallback moot on a system without pidfds.

The fallback if there's no pidfd shall therefore be what we used to do:
relying on the write end of a pipe being closed when the parent dies.

This has some sharp edges, which is why a pidfd will still be used
preferentially.