net: name interfaces optimistically on creation
10a41663 ("nic: rename interfaces after they're created") was initially
written because some older kernels would fail to create an interface in
a target network namespace if an interface with the same name existed in
the original network namespace. This is fixed in newer kernels.

This commit shuffles the code around to benefit from that fix. Now, we
optimistically assume we can create the network interface with the right
name. However, if the link creation fails with EEXIST, we make note of
it and then create the link with a default name, and try a bit later to
rename it in a separate message.